// frontend/lib/generated-api/models/ReproducibilityCategory.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Analyst API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 *
 * @export
 */
export const ReproducibilityCategory = {
  FullyReproducible: 'fully_reproducible',
  ReproducibleWithWebSearch: 'reproducible_with_web_search',
  ReproducibleWithExternalUploads: 'reproducible_with_external_uploads',
  NotReproducible: 'not_reproducible',
} as const;
export type ReproducibilityCategory = (typeof ReproducibilityCategory)[keyof typeof ReproducibilityCategory];

export function instanceOfReproducibilityCategory(value: any): boolean {
  for (const key in ReproducibilityCategory) {
    if (Object.prototype.hasOwnProperty.call(ReproducibilityCategory, key)) {
      if (ReproducibilityCategory[key as keyof typeof ReproducibilityCategory] === value) {
        return true;
      }
    }
  }
  return false;
}

export function ReproducibilityCategoryFromJSON(json: any): ReproducibilityCategory {
  return ReproducibilityCategoryFromJSONTyped(json, false);
}

export function ReproducibilityCategoryFromJSONTyped(json: any, ignoreDiscriminator: boolean): ReproducibilityCategory {
  return json as ReproducibilityCategory;
}

export function ReproducibilityCategoryToJSON(value?: ReproducibilityCategory | null): any {
  return value as any;
}

export function ReproducibilityCategoryToJSONTyped(value: any, ignoreDiscriminator: boolean): ReproducibilityCategory {
  return value as ReproducibilityCategory;
}